Implementing repository pattern for monthly billings (#29)

+++ src/Domain/Exceptions/InvalidYearException.cs
@@ -2,9 +2,9 @@ namespace Domain.Exceptions;
 
 public sealed class InvalidYearException : SmugetException
 {
-    public ushort Value { get; }
+    public int Value { get; }
 
-    public InvalidYearException(ushort value)
+    public InvalidYearException(int value)
         : base($"Year must be an integer number greater than 1900. Passed: {value}.")
     {
         Value = value;
